I think everything is in compliance except...
I did not have a designated container after re-reading the readme.  I was originally thinking wrapping it in a div was putting it in a container.

After looking at http://redux.js.org/ I realize that is not the case.

Suppose wrapping my head around that is the next step.