Complete Todo set up with Redux
I went through and set up the application with the simple functionality to add an item to the Todo list. The set up of Redux wasn't too difficult. The problem came when trying to figure out the best way to handle the Store. Going between the different aspects of Redux, I was losing track of what type of Object the list of Todos was stored in. Once I simplified that, everything worked fine.